--- a/backend/groth16/bls12-377/setup.go
+++ b/backend/groth16/bls12-377/setup.go
@@ -103,7 +103,7 @@ func Setup(r1cs *cs.R1CS, pk *ProvingKey, vk *VerifyingKey) error {
 	}
 
 	// Setting group for fft
-	domain := fft.NewDomain(uint64(len(r1cs.Constraints)))
+	domain := fft.NewDomain(uint64(r1cs.GetNbConstraints()))
 
 	// samples toxic waste
 	toxicWaste, err := sampleToxicWaste()
@@ -334,7 +334,7 @@ func setupABC(r1cs *cs.R1CS, domain *fft.Domain, toxicWaste toxicWaste) (A []fr.
 	var w fr.Element
 	w.Set(&domain.Generator)
 	wi := fr.One()
-	t := make([]fr.Element, len(r1cs.Constraints)+1)
+	t := make([]fr.Element, r1cs.GetNbConstraints()+1)
 	for i := 0; i < len(t); i++ {
 		t[i].Sub(&toxicWaste.t, &wi)
 		wi.Mul(&wi, &w) // TODO this is already pre computed in fft.Domain
@@ -378,23 +378,36 @@ func setupABC(r1cs *cs.R1CS, domain *fft.Domain, toxicWaste toxicWaste) (A []fr.
 	// for each term appearing in the linear expression,
 	// we compute term.Coefficient * L, and cumulate it in
 	// A, B or C at the index of the variable
-	for i, c := range r1cs.Constraints {
 
-		for _, t := range c.L {
-			accumulate(&A[t.WireID()], t, &L)
-		}
-		for _, t := range c.R {
-			accumulate(&B[t.WireID()], t, &L)
-		}
-		for _, t := range c.O {
-			accumulate(&C[t.WireID()], t, &L)
-		}
+	j := 0
+	var c constraint.R1C
+	for _, inst := range r1cs.Instructions {
+		blueprint := r1cs.Blueprints[inst.BlueprintID]
+		if bc, ok := blueprint.(constraint.BlueprintR1C); ok {
+			calldata := r1cs.GetCallData(inst)
+			bc.DecompressR1C(&c, calldata)
 
-		// Li+1 = w*Li*(t-w^i)/(t-w^(i+1))
-		L.Mul(&L, &w)
-		L.Mul(&L, &t[i])
-		L.Mul(&L, &tInv[i+1])
+			for _, t := range c.L {
+				accumulate(&A[t.WireID()], t, &L)
+			}
+			for _, t := range c.R {
+				accumulate(&B[t.WireID()], t, &L)
+			}
+			for _, t := range c.O {
+				accumulate(&C[t.WireID()], t, &L)
+			}
+
+			// Li+1 = w*Li*(t-w^i)/(t-w^(i+1))
+			L.Mul(&L, &w)
+			L.Mul(&L, &t[j])
+			L.Mul(&L, &tInv[j+1])
+
+			j++
+		} else {
+			// TODO @gbotrel blocks
+		}
 	}
+
 	return
 
 }
@@ -448,7 +461,7 @@ func sampleToxicWaste() (toxicWaste, error) {
 func DummySetup(r1cs *cs.R1CS, pk *ProvingKey) error {
 	// get R1CS nb constraints, wires and public/private inputs
 	nbWires := r1cs.NbInternalVariables + r1cs.GetNbPublicVariables() + r1cs.GetNbSecretVariables()
-	nbConstraints := len(r1cs.Constraints)
+	nbConstraints := r1cs.GetNbConstraints()
 
 	// Setting group for fft
 	domain := fft.NewDomain(uint64(nbConstraints))
@@ -526,14 +539,28 @@ func dummyInfinityCount(r1cs *cs.R1CS) (nbZeroesA, nbZeroesB int) {
 
 	A := make([]bool, nbWires)
 	B := make([]bool, nbWires)
-	for _, c := range r1cs.Constraints {
-		for _, t := range c.L {
-			A[t.WireID()] = true
-		}
-		for _, t := range c.R {
-			B[t.WireID()] = true
+
+	j := 0
+	var c constraint.R1C
+	for _, inst := range r1cs.Instructions {
+		blueprint := r1cs.Blueprints[inst.BlueprintID]
+		if bc, ok := blueprint.(constraint.BlueprintR1C); ok {
+			calldata := r1cs.GetCallData(inst)
+			bc.DecompressR1C(&c, calldata)
+
+			for _, t := range c.L {
+				A[t.WireID()] = true
+			}
+			for _, t := range c.R {
+				B[t.WireID()] = true
+			}
+
+			j++
+		} else {
+			continue // TODO @gbotrel handle blocks
 		}
 	}
+
 	for i := 0; i < nbWires; i++ {
 		if !A[i] {
 			nbZeroesA++