@gbotrel 2023-04-19
note: trying to have R1C and Hint blueprint implement solve to recover from solver perf regression.